requirements.txt
Flask==3.1.2
flask-cors==6.0.1
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
numpy==2.2.6
Pillow==11.1.0
python-dotenv==1.2.1
click==8.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
Werkzeug==3.1.3
gunicorn==23.0.0
bcrypt==4.1.2
requests==2.31.0
PyJWT==2.8.0
Flask-Limiter==3.5.0
stripe>=13.0.0
